testing prettier-version on cjs and esm
