+++ package.json
@@ -4,8 +4,8 @@
   "description": "A Prettier Plugin for automatically formatting your Solidity code.",
   "type": "module",
   "main": "./src/index.js",
-  "browser": "./dist/standalone.js",
-  "unpkg": "./dist/standalone.js",
+  "browser": "./dist/standalone.cjs",
+  "unpkg": "./dist/standalone.cjs",
   "scripts": {
     "build": "webpack --env production",
     "build:dev": "webpack --env development",
@@ -14,8 +14,8 @@
     "lint:fix": "npm run eslint -- --fix && npm run prettier -- --write",
     "prepublishOnly": "npm run build",
     "prettier": "prettier './*.{js,json,md,yml}' '{src,tests}/**/*.js'",
-    "test": "jest",
-    "test:all": "cross-env FULL_TEST=1 c8 jest",
+    "test": "NODE_OPTIONS=--loader=esmock jest",
+    "test:all": "cross-env FULL_TEST=1 NODE_OPTIONS=--loader=esmock c8 jest",
     "test:standalone": "cross-env TEST_STANDALONE=1 FULL_TEST=1 jest"
   },
   "files": [
@@ -87,6 +87,7 @@
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
     "esm-utils": "^4.1.2",
+    "esmock": "^2.3.1",
     "exports-loader": "^4.0.0",
     "jest": "^29.6.1",
     "jest-light-runner": "^0.5.0",
